+++ spec/email_inquire_spec.rb
@@ -38,4 +38,40 @@ RSpec.describe EmailInquire do
       expect(described_class.custom_invalid_domains).to eq(Set.new)
     end
   end
+
+  describe ".custom_valid_domains" do
+    context "when untouched" do
+      it "returns a set" do
+        expect(described_class.custom_valid_domains).to be_a(Set)
+      end
+    end
+  end
+
+  describe ".custom_valid_domains=" do
+    it "accepts nil" do
+      described_class.custom_valid_domains = nil
+
+      expect(described_class.custom_valid_domains).to eq(Set.new)
+    end
+
+    it "accepts an Array" do
+      described_class.custom_valid_domains = %w[domain1.com domain2.com]
+
+      expect(described_class.custom_valid_domains).to eq(Set.new(%w[domain1.com domain2.com]))
+    end
+
+    it "accepts a Set" do
+      described_class.custom_valid_domains = Set.new(%w[domain1.com domain2.com])
+
+      expect(described_class.custom_valid_domains).to eq(Set.new(%w[domain1.com domain2.com]))
+    end
+
+    it "raises for unsupported types" do
+      expect {
+        described_class.custom_valid_domains = "domain1.com"
+      }.to raise_error(ArgumentError, "Unsupported type in `custom_valid_domains=`")
+
+      expect(described_class.custom_valid_domains).to eq(Set.new)
+    end
+  end
 end